small ajustment and removed the SLC because of it being useless (in Paragraph).
also added pending in paragraph, this is because there mught not be space in the arrangement or whatever the paragraph is for.

Added the Arrangement container which is for the Arrangement page. It is Atm using divs and such. but I want to implement Grids!

PS: I know Paragraph isnt a suitible name for the component.

SLC: stateless component